templates/cosmosdb-project/python-cosmos/requirements.txt
# Azure CosmosDB Python Development Requirements

# CosmosDB SDK
azure-cosmos>=4.5.0
azure-identity>=1.14.0

# Development and testing
python-dotenv>=1.0.0
pytest>=7.4.0
requests>=2.31.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Logging and monitoring
structlog>=23.1.0